backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
asyncpg==0.29.0
aiohttp==3.9.5
python-dotenv==1.0.1
